+++ shared/lib-bie-kafka/src/main/java/gov/va/vro/model/biekafka/ContentionEvent.java
@@ -1,5 +1,7 @@
 package gov.va.vro.model.biekafka;
 
+import java.util.Arrays;
+
 public enum ContentionEvent {
   CONTENTION_ASSOCIATED_TO_CLAIM("CONTENTION_BIE_CONTENTION_ASSOCIATED_TO_CLAIM_V02"),
   CONTENTION_UPDATED("CONTENTION_BIE_CONTENTION_UPDATED_V02"),
@@ -15,5 +17,20 @@ public enum ContentionEvent {
 
   public String getTopicName() {
     return topicName;
+  }
+
+  public static ContentionEvent mapTopicToEvent(String topic) {
+    // remove first word prefix from topic seperated by _
+    String noPrefixTopic = topic.substring(topic.indexOf("_") + 1);
+
+    return Arrays.stream(ContentionEvent.values())
+        .filter(event -> event.getTopicName().equals(noPrefixTopic))
+        .findFirst()
+        .orElseThrow(() -> new IllegalArgumentException("Unrecognized topic: " + noPrefixTopic));
+  }
+
+  public static String generateRabbitMQChannelName(String topic) {
+    return String.format(
+        "bie-events-%s", mapTopicToEvent(topic).toString().toLowerCase().replace("_", "-"));
   }
 }
