+++ svc-bie-kafka/src/integrationTest/java/gov/va/vro/services/bie/IntegrationTestConfig.java
@@ -1,7 +1,6 @@
 package gov.va.vro.services.bie;
 
-import static gov.va.vro.services.bie.service.kafka.MessageHelper.generateRabbitMQChannelName;
-
+import gov.va.vro.model.biekafka.ContentionEvent;
 import gov.va.vro.services.bie.config.BieProperties;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -35,7 +34,7 @@ public class IntegrationTestConfig {
   }
 
   String mqExchangeName() {
-    return generateRabbitMQChannelName(kafkaTopic());
+    return ContentionEvent.generateRabbitMQChannelName(kafkaTopic());
   }
 
   @Bean
